pyproject.toml
[tool.poetry]
name = "criptam"
version = "1.0"
description = "iOS firmware key decrypter"
authors = ["m1stadev <adamhamdi31@gmail.com>"]
maintainers = ["m1stadev <adamhamdi31@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"
repository = "https://github.com/m1stadev/Criptam"
keywords = ["ios", "jailbreak", "iboot", "img4"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.8",
    "Topic :: Utilities"
]

[tool.poetry.scripts]
criptam = "criptam.__main__:main"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
ipwndfu = "~2.0.0b5"
pyimg4 = "^0.7"
pyusb = "^1.2.1"
remotezip = "^0.9.3"
requests = "^2.27.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"


[tool.black]
skip-string-normalization = true

[tool.isort]
profile = "black"
src_paths = ["criptam"]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/m1stadev/Criptam/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
